requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
distro==1.8.0
dnspython==2.4.2
exceptiongroup==1.2.0
fastapi==0.105.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
httpx-sse==0.3.1
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.1
langserve==0.0.36
langsmith==0.0.70
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
openai==1.4.0
orjson==3.9.10
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pinecone-client==2.2.4
protobuf==4.25.1
pyarrow==14.0.1
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
sse-starlette==1.8.2
starlette==0.27.0
streamlit==1.29.0
streamlit-chat==0.1.1
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
transformers==4.36.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.24.0.post1
validators==0.22.0
yarl==1.9.4
zipp==3.17.0
